backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pytesseract==0.3.10
Pillow==10.4.0
numpy==2.0.1
opencv-python==4.10.0.84
scikit-image==0.24.0
orjson==3.10.7


